first upload of some older scripts
First upload of some older scripts around for the purpose of desktop configuration automation. Most of these are written to use much older versions of PowerShell so use with caution. Additionally, there may be actual PowerShell commands to complete some of the work that needed to be done previously using other methods. For example, Rename-Computer will also rename a machine whereas before you had to use wmi commands to complete a machine rename.